+++ F1G121050 - Dwi Irman Saputra/MENCARI LUAS LINGKARAN.cpp
@@ -1,0 +1,16 @@
+#include <iostream>
+using namespace std;
+
+int main (){
+	// Dwi irman saputra_F1G121050
+	// Mencari luas lingkaran
+	
+	float r, luas, phi = 3.14;
+	cout<<"Masukan nilai r = ";
+	cin>> r;
+	
+	luas= phi*r*r;
+	cout<<"Hasil luas lingkaran adalah = "<<luas<<" cm";
+	
+	return 0;
+}
