+++ F1G121050 - Dwi Irman Saputra/MENCARI VOLUME KUBUS.cpp
@@ -1,0 +1,17 @@
+#include <iostream>
+using namespace std;
+
+int main (){
+	// Dwi irman saputra_FIGI21050
+	// Mencari volume kubus
+	
+	float a,hasil;
+	cout<<"Masukan nilai sisi = ";
+	cin>> a;
+	
+	hasil=a*a*a;
+	cout<<"Hasil prhitungan mencari volume kubus adalah = "<<hasil<<" cm";
+	
+	return 0;
+	
+}
